+++ src/word_vector.hpp
@@ -24,7 +24,7 @@ limitations under the License.
 
 namespace resembla {
 
-template<typename string_type, typename value_type = double, typename id_type = long>
+template<typename string_type, typename value_type = float, typename id_type = long>
 struct WordVector
 {
     id_type id;
